requirements.txt
arango-orm==0.7.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.0.12
DateTime==5.1
Django==4.1.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
future==0.18.3
idna==3.4
marshmallow==3.19.0
packaging==23.0
pyArango==2.0.1
pydantic==1.10.7
PyJWT==2.6.0
python-arango==7.5.7
pytz==2023.3
replicate==0.5.2
requests==2.27.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
zope.interface==6.0
